refactor: Group 에 속해있는 Participant 를 별도의 패키지로 분리
