+++ backend/src/main/java/com/woowacourse/momo/participant/service/GroupParticipantService.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.group.service;
+package com.woowacourse.momo.participant.service;
 
 import java.util.List;
